# io/rest/app/controllers.go
// Code generated by goagen v1.3.0, DO NOT EDIT.
//
// API "ocg": Application Controllers
//
// Command:
// $ goagen
// --design=github.com/aweisser/ocg-poc/io/rest/design
// --out=$(GOPATH)\src\github.com\aweisser\ocg-poc\io\rest\
// --version=v1.1.0-dirty

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
	"regexp"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// MemberController is the controller interface for the Member actions.
type MemberController interface {
	goa.Muxer
	Show(*ShowMemberContext) error
}

// MountMemberController "mounts" a Member resource controller on the given service.
func MountMemberController(service *goa.Service, ctrl MemberController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/members/:memberID", ctrl.MuxHandler("preflight", handleMemberOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewShowMemberContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Show(rctx)
	}
	h = handleMemberOrigin(h)
	service.Mux.Handle("GET", "/api/members/:memberID", ctrl.MuxHandler("show", h, nil))
	service.LogInfo("mount", "ctrl", "Member", "action", "Show", "route", "GET /api/members/:memberID")
}

// handleMemberOrigin applies the CORS response headers corresponding to the origin.
func handleMemberOrigin(h goa.Handler) goa.Handler {
	spec0 := regexp.MustCompile("[localhost]")

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOriginRegexp(origin, spec0) {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
				rw.Header().Set("Access-Control-Allow-Headers", "X-Request-Id, Content-Type, Authorization, If-None-Match, If-Modified-Since")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}