# io/rest/swagger/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  GenericData:
    example:
      id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
      links:
        meta:
          Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
        self: Voluptatum consequatur.
      type: Illo temporibus nobis voluptate sapiente.
    properties:
      id:
        description: UUID of the object
        example: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
        type: string
      links:
        $ref: '#/definitions/GenericLinks'
      type:
        example: Illo temporibus nobis voluptate sapiente.
        type: string
    title: GenericData
    type: object
  GenericLinks:
    example:
      meta:
        Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
      self: Voluptatum consequatur.
    properties:
      meta:
        additionalProperties: true
        example:
          Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
        type: object
      self:
        example: Voluptatum consequatur.
        type: string
    title: GenericLinks
    type: object
  GenericLinksForMember:
    example:
      meta:
        Et provident vel qui consequuntur voluptatem voluptatem.: true
      self: Numquam maiores rerum nihil.
    properties:
      meta:
        additionalProperties: true
        example:
          Et provident vel qui consequuntur voluptatem voluptatem.: true
        type: object
      self:
        example: Numquam maiores rerum nihil.
        type: string
    title: GenericLinksForMember
    type: object
  JSONAPIError:
    description: |-
      Error objects provide additional information about problems encountered while
      performing an operation. Error objects MUST be returned as an array keyed by errors in the
      top level of a JSON API document.

      See. also http://jsonapi.org/format/#error-objects.
    example:
      code: Et dolor earum saepe recusandae ipsum.
      detail: Eos veritatis eius.
      id: Porro id tenetur deleniti dolorem.
      links:
        Sequi voluptas.:
          href: http://example.com/articles/1/comments
          meta:
            Voluptatem quia ut saepe quia.: 2000-08-31T08:02:30Z
      meta:
        Et quam numquam ut distinctio veritatis.: true
      source:
        Sapiente aut quasi sit.: 2008-05-03T15:36:04Z
      status: Maiores maiores id est et cum ipsum.
      title: Sed recusandae similique.
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: Et dolor earum saepe recusandae ipsum.
        type: string
      detail:
        description: |-
          a human-readable explanation specific to this occurrence of the problem.
          Like title, this field’s value can be localized.
        example: Eos veritatis eius.
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: Porro id tenetur deleniti dolorem.
        type: string
      links:
        additionalProperties: true
        description: |-
          a links object containing the following members:
          * about: a link that leads to further details about this particular occurrence of the problem.
        example:
          Sequi voluptas.:
            href: http://example.com/articles/1/comments
            meta:
              Voluptatem quia ut saepe quia.: 2000-08-31T08:02:30Z
        type: object
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error
        example:
          Et quam numquam ut distinctio veritatis.: true
        type: object
      source:
        additionalProperties: true
        description: |-
          an object containing references to the source of the error,
          optionally including any of the following members

          * pointer: a JSON Pointer [RFC6901] to the associated entity in the request document [e.g. "/data" for a primary data object,
                     or "/data/attributes/title" for a specific attribute].
          * parameter: a string indicating which URI query parameter caused the error.
        example:
          Sapiente aut quasi sit.: 2008-05-03T15:36:04Z
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: Maiores maiores id est et cum ipsum.
        type: string
      title:
        description: |-
          a short, human-readable summary of the problem that SHOULD NOT
          change from occurrence to occurrence of the problem, except for purposes of localization.
        example: Sed recusandae similique.
        type: string
    required:
    - detail
    title: JSONAPIError
    type: object
  JSONAPIErrors:
    description: JSONAPIErrors media type (default view)
    example:
      errors:
      - code: Et dolor earum saepe recusandae ipsum.
        detail: Eos veritatis eius.
        id: Porro id tenetur deleniti dolorem.
        links:
          Sequi voluptas.:
            href: http://example.com/articles/1/comments
            meta:
              Voluptatem quia ut saepe quia.: 2000-08-31T08:02:30Z
        meta:
          Et quam numquam ut distinctio veritatis.: true
        source:
          Sapiente aut quasi sit.: 2008-05-03T15:36:04Z
        status: Maiores maiores id est et cum ipsum.
        title: Sed recusandae similique.
      - code: Et dolor earum saepe recusandae ipsum.
        detail: Eos veritatis eius.
        id: Porro id tenetur deleniti dolorem.
        links:
          Sequi voluptas.:
            href: http://example.com/articles/1/comments
            meta:
              Voluptatem quia ut saepe quia.: 2000-08-31T08:02:30Z
        meta:
          Et quam numquam ut distinctio veritatis.: true
        source:
          Sapiente aut quasi sit.: 2008-05-03T15:36:04Z
        status: Maiores maiores id est et cum ipsum.
        title: Sed recusandae similique.
    properties:
      errors:
        example:
        - code: Et dolor earum saepe recusandae ipsum.
          detail: Eos veritatis eius.
          id: Porro id tenetur deleniti dolorem.
          links:
            Sequi voluptas.:
              href: http://example.com/articles/1/comments
              meta:
                Voluptatem quia ut saepe quia.: 2000-08-31T08:02:30Z
          meta:
            Et quam numquam ut distinctio veritatis.: true
          source:
            Sapiente aut quasi sit.: 2008-05-03T15:36:04Z
          status: Maiores maiores id est et cum ipsum.
          title: Sed recusandae similique.
        - code: Et dolor earum saepe recusandae ipsum.
          detail: Eos veritatis eius.
          id: Porro id tenetur deleniti dolorem.
          links:
            Sequi voluptas.:
              href: http://example.com/articles/1/comments
              meta:
                Voluptatem quia ut saepe quia.: 2000-08-31T08:02:30Z
          meta:
            Et quam numquam ut distinctio veritatis.: true
          source:
            Sapiente aut quasi sit.: 2008-05-03T15:36:04Z
          status: Maiores maiores id est et cum ipsum.
          title: Sed recusandae similique.
        items:
          $ref: '#/definitions/JSONAPIError'
        type: array
    required:
    - errors
    title: 'Mediatype identifier: application/vnd.jsonapierrors+json; view=default'
    type: object
  Member:
    example:
      attributes:
        name: Johndoe42
      id: 4711xyz
      links:
        meta:
          Et provident vel qui consequuntur voluptatem voluptatem.: true
        self: Numquam maiores rerum nihil.
      relationships:
        accounts:
          data:
            id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
            links:
              meta:
                Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
              self: Voluptatum consequatur.
            type: Illo temporibus nobis voluptate sapiente.
          links:
            meta:
              Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
            self: Voluptatum consequatur.
          meta:
            Iure deleniti dignissimos dolor reiciendis explicabo aut.: 1.2256759e+18
      type: members
    properties:
      attributes:
        $ref: '#/definitions/memberAttributes'
      id:
        description: The unique ID of the member
        example: 4711xyz
        type: string
      links:
        $ref: '#/definitions/GenericLinksForMember'
      relationships:
        $ref: '#/definitions/MemberRelationships'
      type:
        description: The type of the related resource
        enum:
        - members
        example: members
        type: string
    required:
    - type
    - attributes
    title: Member
    type: object
  MemberRelationships:
    example:
      accounts:
        data:
          id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
          links:
            meta:
              Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
            self: Voluptatum consequatur.
          type: Illo temporibus nobis voluptate sapiente.
        links:
          meta:
            Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
          self: Voluptatum consequatur.
        meta:
          Iure deleniti dignissimos dolor reiciendis explicabo aut.: 1.2256759e+18
    properties:
      accounts:
        $ref: '#/definitions/RelationGeneric'
    title: MemberRelationships
    type: object
  MemberSingle:
    description: Holds a single response to a member request (default view)
    example:
      data:
        attributes:
          name: Johndoe42
        id: 4711xyz
        links:
          meta:
            Et provident vel qui consequuntur voluptatem voluptatem.: true
          self: Numquam maiores rerum nihil.
        relationships:
          accounts:
            data:
              id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
              links:
                meta:
                  Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
                self: Voluptatum consequatur.
              type: Illo temporibus nobis voluptate sapiente.
            links:
              meta:
                Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
              self: Voluptatum consequatur.
            meta:
              Iure deleniti dignissimos dolor reiciendis explicabo aut.: 1.2256759e+18
        type: members
      included:
      - false
      - false
    properties:
      data:
        $ref: '#/definitions/Member'
      included:
        description: An array of mixed types
        example:
        - false
        - false
        items:
          example: false
        type: array
    required:
    - data
    title: 'Mediatype identifier: application/vnd.member+json; view=default'
    type: object
  RelationGeneric:
    example:
      data:
        id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
        links:
          meta:
            Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
          self: Voluptatum consequatur.
        type: Illo temporibus nobis voluptate sapiente.
      links:
        meta:
          Enim et possimus mollitia.: 36bfc211-9a28-46e2-9853-3f5adc06b87f
        self: Voluptatum consequatur.
      meta:
        Iure deleniti dignissimos dolor reiciendis explicabo aut.: 1.2256759e+18
    properties:
      data:
        $ref: '#/definitions/GenericData'
      links:
        $ref: '#/definitions/GenericLinks'
      meta:
        additionalProperties: true
        example:
          Iure deleniti dignissimos dolor reiciendis explicabo aut.: 1.2256759e+18
        type: object
    title: RelationGeneric
    type: object
  memberAttributes:
    example:
      name: Johndoe42
    properties:
      name:
        description: Name of the member
        example: Johndoe42
        maxLength: 62
        minLength: 1
        pattern: ^[^_|-].*
        type: string
    required:
    - name
    title: memberAttributes
    type: object
host: localhost
info:
  description: OCG Rest API
  license:
    name: GNU GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.txt
  title: OCG Rest API
  version: "1.0"
paths:
  /members/{memberID}:
    get:
      description: Retrieve member (as JSONAPI) for the given ID.
      operationId: member#show
      parameters:
      - description: ID of the member
        in: path
        name: memberID
        required: true
        type: string
      - in: header
        name: If-Modified-Since
        required: false
        type: string
      - in: header
        name: If-None-Match
        required: false
        type: string
      produces:
      - application/vnd.jsonapierrors+json
      - application/vnd.member+json
      responses:
        "200":
          description: Resource created
          headers:
            Cache-Control:
              type: string
            ETag:
              type: string
            Last-Modified:
              type: string
          schema:
            $ref: '#/definitions/MemberSingle'
        "304":
          description: Not Modified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/JSONAPIErrors'
      schemes:
      - http
      summary: show member
      tags:
      - member
produces:
- application/json
responses:
  NotModified:
    description: Not Modified
  OK:
    description: Resource created
    headers:
      Cache-Control:
        type: string
      ETag:
        type: string
      Last-Modified:
        type: string
schemes:
- http
swagger: "2.0"